Allow Knight to be Placed and then call Tour
Knight is placed by having a click event on the tile that then passes
it index value back up to board. Not sure if this is overcomplicating
matters. I will need to revisit. There are two booleans that we check
if the knight is placed and then if a second location has been set.
If there is a knight but no ending location, we call the tour
function that returns an array of the moves it needs to make. Right
now, we are just console logging, but we need to handle moving the
knight next.